\name{solidauthr-package}
\alias{solidauthr}
\alias{solidauthr-package}
\title{solidauthr: Solid OIDC Client Credentials Authentication}
\description{
Utilities for authenticating against Community Solid Server identity providers
with OAuth client credentials and DPoP, then issuing authenticated requests to
Solid resources.
}
\docType{package}
